LILLIPUT: Fix bug in OC_compWord10804

+++ engines/lilliput/lilliput.h
@@ -337,8 +337,8 @@ public:
 	byte _byte12FCE;
 	byte _byte129A0;
 	byte _numCharactersToDisplay;
-	byte _byte16C9F;
-	int _word10804;
+	byte _byte16C9F; 
+	int16 _word10804;
 	bool _shouldQuit;
